+++ frontends/PyCDE/integration_test/esi.py
@@ -94,7 +94,3 @@ s.build_api("python")
 
 sys.path.append(sys.argv[1])
 from esi_rt.ESILoopback import top
-
-import IPython
-
-IPython.embed(locals=locals())
